Mapview stops following user if user drags map
pursuant to the ‘react-native-maps’ docs, (https://github.com/airbnb/react-native-maps/blob/master/docs/mapview.md):

“ onPanDrag - callback that is called when user presses and drags the map. NOTE: for iOS scrollEnabled should be set to false to trigger the event “

thus, we need to set the region via a prop passed into the map and track the state of both scrollEnabled and shouldMapFollowUser.

furthermore, since the docs say ‘followsUserLocation’ only works on iOS, we follow the user by continuously passing the map a new region as a prop.

WBMapView.js is passed a callback from MainScreen.js which is called when onPanDrag is called. That disables isScrollEnabled and Since this is called multiple times per drag, the user will never notice that the first onPanDrag didn’t actually move the map.