Bugfix: map no longer recenters on compass heading
since MainScreen.js’ render function was passing in the this.state.wbMapRegion, whenever the map would re-render it would re-render with that region. This would happen when the compass adjusted (which was basically each moment when a user is holding their phone) and would also happen when the user would drag up the drag bar and resize the map.

so fix we simply track the current state of the map region, so when that re-render happens, we can use that current state in the render.