All ignite gloat and boilerplate removed
initially, I used ‘ignite’ to generate much of the code which got me started and provided something to learn react-native from. Much of it was not used, yet is slowing build times, and making this project difficult to read.  This commit removes all of that.

for reference: see, https://github.com/infinitered/ignite

ignite seems better suited for larger projects.

NOTE1: For this this commit, I generated a new react-native project from scratch, and brought in only the code and dependencies that are being used. Then I deleted everything from this project, and copied that new project into this project directory. Some functionality may be lost, particularly on Android builds.

NOTE2: Commit messages and blames may be lost in most files before this commit. I tried to keep them on MainScreen.js. However, you may need to roll back before this commit if you need to see a commit message.